private-socket-master/requirements.txt
python-socketio==5.4.1
eventlet==0.32.0
